move sample
move so we can have multiple samples.  Fix existing sample too.
