+++ GooseHtml.Samples/Mvc/Properties/launchSettings.json
