+++ GooseHtml.Samples/Mvc/appsettings.Development.json
