+++ GooseHtml.Samples/Mvc/appsettings.json
